Update blackjack_me_draft_5(implementing_counting).py
I can calculate the frequency of true counts and the decks_remaining divisor is now by halves. I plan to make it by quarters for double deck so that that frequency lines up properly with CVCX but so far six deck count frequency calculations are behaving properly which is awesome.